health cehck move andpoint + add xml. to refactor
