+++ src/AppBundle/Resources/views/Manage/availability.html.twig
