+++ src/AppBundle/Resources/views/Manage/health-check.xml.twig
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<pingdom_http_custom_check>
+  <status>{{ status }}</status>
+  <response_time>{{ time | round(3) }}</response_time>
+</pingdom_http_custom_check>
