+++ tests/behat/features/01-admin.feature
@@ -1,10 +1,10 @@
 Feature: admin
     
     @deputy @admin
-    Scenario: reset behat data before starting
-        Given I am on "/health-check"
+    Scenario: check app before starting
+        Given I am on "/manage/availability/health-check.xml"
         Then the response status code should be 200
-        And I reset the behat data
+        #And I reset the behat data # not implemented yet. run phing behat on api and then on the client to prepare
 
     @deputy
     Scenario: login and add deputy user
